feed cache

+++ ImageConverter.class.php
@@ -15,12 +15,15 @@ class ImageConverter
     {
         $name = $this->folder . DIRECTORY_SEPARATOR . md5($file);
         $outfile = dirname($file) . DIRECTORY_SEPARATOR . basename($file, ".img") . '.jpg';
-        file_put_contents($this->base . DIRECTORY_SEPARATOR . $name . '.convert', json_encode([
-            'file' => $file,
-            'outfile' => $outfile,
-            'width' => 700,
-            'height' => 700
-        ]));
+
+        if (!file_exists($this->base . $outfile)) {
+            file_put_contents($this->base . DIRECTORY_SEPARATOR . $name . '.convert', json_encode([
+                'file' => $file,
+                'outfile' => $outfile,
+                'width' => 700,
+                'height' => 700
+            ]));
+        }
 
         return $outfile;
     }
